Initial DRY refactor
